TPC: Adding loading of distortions maps from CCDB for MC
Distortions from pre-calculated maps can be loaded from CCDB with the option
--tpc-simulate-distortions in the o2-sim-digitizer-workflow:

o2-sim-digitizer-workflow $ARGS_ALL -b --tpc-lanes 1 --onlyDet TPC --tpc-simulate-distortions

+++ Detectors/TPC/simulation/include/TPCSimulation/Digitizer.h
@@ -53,7 +53,7 @@ enum class SCDistortionType : int;
 class Digitizer
 {
  public:
-  using SC = SpaceCharge<double>;
+  using SC = SpaceCharge<float>;
 
   /// Default constructor
   Digitizer();
